Added blackbox test for all dds-types-test (#3)
* Created Typed Tests


* Added submodule


* Added blackbox typed test


* Fix for empty filename in log callback


* Named typed tests. Disabled stdout log.


* Reworked ddsEnabler blackbox tests.


* Removed old file. Applied suggestions.


* Rebased.


* Fixed bug disabling stdout log.


* Changed publishedTime to int64.


* Added check for empty callbacks. Removed missing header.


* Added tests for history.


* Included cstdint for callbacks.


* Remove comment


---------

+++ ddsenabler_participants/include/ddsenabler_participants/CBCallbacks.hpp
@@ -18,6 +18,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 namespace eprosima {
 namespace ddsenabler {
 namespace participants {
@@ -47,7 +49,7 @@ typedef void (*DdsNotification)(
         const char* typeName,
         const char* topicName,
         const char* json,
-        double publishTime);
+        int64_t publishTime);
 
 
 } /* namespace participants */
